# backend/api/openapi.yaml
...
openapi: 3.1.0
...
info:
  title: Foraged goods marketplace API
  version: 0.1.0
  # "|" keeps the newlines - the description renders as Markdown in the UI.
  description: |
    A marketplace for foraged goods — berries, mushrooms, herbs. Sellers post
    listings, buyers browse, message them and place orders. Payments are
    simulated; no money changes hands.

    ## Authentication

    Two tokens, with opposite properties.

    The **access token** is a signed JWT sent on every request:

        Authorization: Bearer <access_token>

    It is verified without touching the database, which also means it cannot be
    revoked - so it is short-lived (15 minutes by default, `ACCESS_TOKEN_TTL`).

    The **refresh token** is an opaque value in an HttpOnly `refresh_token`
    cookie, set by signup and login and valid for 7 days. `POST /auth/refresh`
    exchanges it for a new access token and rotates it. Only its hash is
    stored, so a database leak cannot resume a session.

    When an access token expires, the 401 carries
    `WWW-Authenticate: Bearer error="invalid_token"` - the signal to refresh
    rather than to show a login screen. `GET /auth/me` returns the identity
    behind a token that is still valid, for restoring a session on page load.

    Public endpoints (browsing listings, search, images) work without one.

    ## Ids

    Every id is a **uuid v7**: the leading 48 bits are a millisecond timestamp,
...
    something to parse.

    ## Errors

...

...

...
servers:
...

...
security:
  - bearerAuth: []
  - apiKeyAuth: []

# "components" is the reusable-definitions section. Nothing in here appears in
# the docs on its own - it only shows up where something $refs it.
components:
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        A key from `POST /me/api-keys`, for machine clients. It acts as its
        owner - same permissions, same ownership checks - and is rate limited
        per key rather than per IP.

        Keys cannot manage keys: creating, listing and revoking need a session,
        so a leaked key cannot mint a replacement and survive its own
        revocation.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT # documentation only; no tool verifies the format
      description: |
        Access token issued by signup or login. Expired and missing tokens are
        both answered with 401.
  schemas:
...
    Error:
      type: object
      required: [error] # "details" is omitempty in Go, so it is not required
      properties:
        error:
          type: string
          description: |
            Safe to show a user on 4xx. On 5xx this is always
            "something went wrong" — the cause is in the server log.
          examples: ["listing not found"]
        details:
          type: object
          # "any key, string value" - the YAML spelling of map[string]string.
          additionalProperties:
            type: string
          description: Reserved for field-level validation errors; absent today.
    ListingImage:
      type: object
      required: [id, url, position]
      properties:
        id:
          type: string
          format: uuid
          examples: ["01a02305-b81c-7dcb-86a0-7f75e33e0af3"]
        url:
          type: string
          description: Path served by the API, e.g. for use in an `img` tag.
          examples: ["/uploads/0a7a0bc6-99f9-42f8-9ba0-c908a9cb3a93.png"]
        position:
          type: integer
          format: int32
          description: Display order, ascending. Unique per listing.
          examples: [0]
...
    CreateOrderInput:
      type: object
      required: [listing_id, quantity]
      properties:
        listing_id:
          type: string
          format: uuid
        quantity:
          type: integer
          format: int32
          exclusiveMinimum: 0
    Order:
      type: object
      required:
        [
          id,
          listing_id,
          listing_title,
          buyer_id,
          seller_id,
          quantity,
          unit_price,
          total_price,
          status,
          seller_handed_over_at,
          buyer_received_at,
          created_at,
          updated_at,
        ]
      properties:
        id:
          type: string
          format: uuid
        listing_id:
          type: string
          format: uuid
        listing_title:
          type: string
          description: |
            Snapshotted when the order was placed, so the order still reads
            correctly after the listing is edited.
        buyer_id:
          type: string
          format: uuid
        seller_id:
          type: string
          format: uuid
        quantity:
          type: integer
          format: int32
        unit_price:
          type: number
          format: double
          description: |
            Price per unit, snapshotted when the order was placed, so the order
            still reads correctly after the listing is repriced.

            Stored as `NUMERIC(10,2)` and converted only on the way out —
            `total_price` is multiplied by Postgres in exact decimal, never in
            Go. Trailing zeroes are not preserved: 18.50 is sent as 18.5, and
            74.00 as 74. Format for display client-side.
          examples: [18.5]
        total_price:
          type: number
          format: double
          description: unit_price × quantity, computed in the database.
          examples: [74]
        status:
          $ref: "#/components/schemas/OrderStatus"
        seller_handed_over_at:
          # "null OR a timestamp" - the Go field is *time.Time.
          type: [string, "null"]
          format: date-time
          description: Set when the seller marked handover; null until then.
        buyer_received_at:
          type: [string, "null"]
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    OrderStatus:
      type: string
      description: |
        The lifecycle: `pending` → `confirmed` → `completed`, or `cancelled`
        from either of the first two.

        `completed` needs BOTH sides: the seller calls `/handover` and the
        buyer calls `/receive`. Whichever comes first only stamps its own
        timestamp and leaves the status at `confirmed`.

        An earlier schema had a `paid` status; migration 007 removed it.
      enum: [pending, confirmed, completed, cancelled]

    # --- chat & settings ------------------------------------------------------
...
    OrderEvent:
      type: object
      description: |
        One entry in an order's history, written in the same transaction as the
        change it describes - so a refused transition leaves no entry, and an
        entry cannot describe something that did not happen.
      required: [id, order_id, actor_id, from_status, to_status, note, created_at]
      properties:
        id:
          type: string
          format: uuid
        order_id:
          type: string
          format: uuid
        actor_id:
          type: [string, "null"]
          format: uuid
          description: |
            Who acted. Null once that account is deleted: the event outlives
            the person, unlike a follow or a session.
        from_status:
          # Deliberately NOT the OrderStatus enum: these are historical values.
          # Migration 007 removed 'paid' from the allowed set, and rows written
          # before that still say it.
          type: [string, "null"]
          description: |
            The status before the change, or null for the event recording the
            order being placed. A plain string rather than the `OrderStatus`
            enum, because old rows may hold statuses no longer in use.
          examples: ["confirmed"]
        to_status:
          type: string
          description: |
            The status after the change. Equal to `from_status` on a handshake
            mark, where a party acted but the order did not move.
          examples: ["completed"]
        note:
          type: [string, "null"]
          description: |
            Set on handshake marks, where `from_status` equals `to_status` and
            the status alone cannot say who acted. Null on ordinary
            transitions.

            A stable code, not display text - `seller_handover` or
            `buyer_receipt` - so a client renders it in its own language.
            Rows already written cannot be re-translated.
          examples: ["seller_handover"]
        created_at:
          type: string
          format: date-time

    # --- profiles --------------------------------------------------------------
...
  responses:
    BadRequest:
      description: The request was malformed or failed validation.
      content:
        application/json: # the media type the body is sent as
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: |
        No token, or one that is expired or invalid. The body is the same in
        every case; `WWW-Authenticate` says which, because a client's response
        differs: refresh, or log in.
      headers:
        WWW-Authenticate:
          description: |
            `Bearer error="invalid_token", error_description="The access token expired"`
            when the token was valid but has aged out - retry through
            `POST /auth/refresh`. A bare `Bearer` for a missing, malformed or
            forged token - the client should log in. A forged token reads as
            absent on purpose.
          schema:
            type: string
          examples:
            expired:
              value: Bearer error="invalid_token", error_description="The access token expired"
            absent:
              value: Bearer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: |
        Authenticated, but not permitted — the resource belongs to someone
        else, or the account does not hold the role the route requires.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: |
        No such resource — also returned when a resource exists but the caller
        is not a participant, so its existence is not revealed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    RateLimited:
      description: |
        Too many requests on this API key. Only key-authenticated requests are
        limited; browser sessions are not.
      headers:
        Retry-After:
          description: Seconds until a request will be accepted again.
          schema:
            type: integer
            examples: [2]
        X-RateLimit-Limit:
          description: |
            The bucket's capacity. Sent on every key-authenticated response,
            not only on a refusal.
          schema:
            type: integer
            examples: [60]
        X-RateLimit-Remaining:
          description: Requests left in the current bucket.
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ServerError:
      description: Something failed on our side; the cause is in the log.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

# Each key is a URL relative to the server above, so "/listings" is really
# http://localhost:8080/api/v1/listings.
paths:
...
  /listings/{id}/images:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [images]
      summary: List a listing's photos
      operationId: getListingImages
      security: []
      description: |
        Ordered by `position`. A listing id that does not exist returns an
        empty array, not a 404.
      responses:
        "200":
          description: The photos, ascending by position.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ListingImage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"

    post:
      tags: [images]
      summary: Upload a photo
      operationId: uploadListingImage
      description: |
        `multipart/form-data` with one file field named `image`. The type is
        decided by sniffing the file's first 512 bytes, not by the filename or
        the client's Content-Type — so a `.png` that is really a PDF is a 415.

        Limits come from the environment: `MAX_UPLOAD_BYTES` (default 5 MiB)
        and `MAX_IMAGES_PER_LISTING` (default 5).
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [image]
              properties:
                image:
                  type: string
                  format: binary # "this field is a file", not text
      responses:
        "201":
          description: |
            Stored. `position` is one past the highest already in use, so a
            gap left by a deleted photo is never reused.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingImage"
        "400":
          description: Not a multipart form, or no `image` field in it.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: This listing already has the maximum number of images.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: The upload exceeded `MAX_UPLOAD_BYTES`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "415":
          description: Not a JPEG, PNG or WebP.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"

  # ---------------------------------------------------------------------------
  # served outside /api/v1 - hence the per-path "servers" override below
  # ---------------------------------------------------------------------------
...
  /listings/{id}/images/{imageID}:
    # Two path parameters. Note the casing: chi's route says {imageID}, so the
    # spec has to say {imageID} too.
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: imageID
        in: path
        required: true
        schema:
          type: string
          format: uuid
    delete:
      tags: [images]
      summary: Delete a photo
      operationId: deleteListingImage
      description: |
        Removes the row and the file. Remaining photos keep their positions;
        gaps are allowed.
      responses:
        "204":
          description: Deleted. No body.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"

  # ---------------------------------------------------------------------------
  # wishlist
  # ---------------------------------------------------------------------------
...
  /orders:
    post:
      tags: [orders]
      summary: Place an order
      operationId: createOrder
      description: |
        Decrements the listing's stock in the same transaction, so two buyers
        racing for the last item cannot both succeed. Starts as `pending`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderInput"
      responses:
        "201":
          description: Order placed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: |
            Bad body, non-positive quantity, or your own listing — you cannot
            buy from yourself.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Not enough stock available.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
    get:
      tags: [orders]
      summary: List your orders
      operationId: getOrders
      description: Every order where you are the buyer OR the seller.
      responses:
        "200":
          description: Your orders.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
  /orders/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [orders]
      summary: Get one order
      operationId: getOrder
      responses:
        "200":
          description: The order.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
  /orders/{id}/events:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [orders]
      summary: An order's history
      operationId: getOrderEvents
      description: |
        Every transition the order made, oldest first: who did what, and when.
        A completed order has five entries - placed, confirmed, both halves of
        the handshake, and the completion.

        The two handshake entries share a `created_at`, because they are
        written in one transaction; they are ordered by `id` behind it.

        Buyer and seller only.
      responses:
        "200":
          description: The timeline, oldest first.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderEvent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: |
            You are not the buyer or the seller. Note orders answer this with
            403 where conversations answer 404 - so an outsider can tell that
            an order id exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
  /orders/{id}/confirm:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [orders]
      summary: Seller accepts the order
      operationId: confirmOrder
      description: "`pending` → `confirmed`. Seller only."
      responses:
        "200":
          description: The order, now `confirmed`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: The order is not `pending`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
  /orders/{id}/handover:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [orders]
      summary: Seller marks the goods handed over
      operationId: handoverOrder
      description: |
        Half of the two-sided handshake. Stamps `seller_handed_over_at`; the
        status only becomes `completed` once the buyer has called `/receive`
        as well. Calling it twice is a 409.
      responses:
        "200":
          description: |
            The order — `completed` if the buyer already confirmed receipt,
            otherwise still `confirmed` with the timestamp now set.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Wrong status, or you already marked handover.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
  /orders/{id}/receive:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [orders]
      summary: Buyer confirms receipt
      operationId: receiveOrder
      description: |
        The other half of the handshake. Stamps `buyer_received_at`; the
        status becomes `completed` only once the seller has handed over too.
      responses:
        "200":
          description: The order, `completed` if both sides have now marked.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Wrong status, or you already confirmed receipt.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
  /orders/{id}/cancel:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [orders]
      summary: Cancel an order
      operationId: cancelOrder
      description: |
        Either side, while the order is `pending` or `confirmed`. The reserved
        stock goes back to the listing. Once either side has marked the
        handshake, cancelling is refused.
      responses:
        "200":
          description: The order, now `cancelled`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Already finished, or handover has started.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"

  # ---------------------------------------------------------------------------
  # chat
  # ---------------------------------------------------------------------------
...